Fix loading bug with FbEvents
Previously if the page was not the first page loaded, the appId would already
have been loaded, so the watch expression would not trigger.
Add a mounted hook that runs if appId is set and non-empty.
Move adding window.fbAsyncInit to a function inside a check if that function
already exists.